Most of MetadataActor is really generic throttling code, move it out
